+++ src/stories/admin/admin.stories.tsx
@@ -6,7 +6,7 @@ import { Admin, AdminContext, AdminUI } from "@/components/admin/admin";
 import { ListGuesser } from "@/components/admin/list-guesser";
 import { ShowGuesser } from "@/components/admin/show-guesser";
 import { EditGuesser } from "@/components/admin/edit-guesser";
-import { bwTheme } from "@/components/admin/bw-theme";
+import { bwTheme } from "@/lib/themes/bw-theme";
 
 export default {
   title: "Application configuration/Admin",